move qheap type definitions to @types folder
