+++ packages/client/tsconfig.browser.json
@@ -5,5 +5,6 @@
   "compilerOptions": {
     "target": "ES2017",
     "outDir": "dist.browser",
+    "typeRoots": ["node_modules/@types", "lib/@types"],
   }
 }
