+++ packages/client/tsconfig.json
@@ -2,5 +2,6 @@
   "extends": "../../config/tsconfig.json",
   "include": ["bin", "lib", "test"],
   "compilerOptions": {
+    "typeRoots": ["node_modules/@types", "lib/@types"],
   },
 }
