+++ packages/client/tsconfig.prod.json
@@ -4,6 +4,7 @@
   "compilerOptions": {
     "outDir": "dist",
     "lib": ["es2018", "dom"],
+    "typeRoots": ["node_modules/@types", "lib/@types"],
   },
   "references": [
     { "path": "../block/tsconfig.prod.json" },
